First commit of map export to dmm.
It's currently an adminsecret that has you pick two turfs and a filename. As is, it saves way more info than needed and only leaves out mobs with ckeys.
The exported maps will be in the same folder as the normal ones and files with the same name are not handled yet.

It's slow and needs optimisation, saving large areas is not recommended. I wouldn't recommend it for use for anything important but it's here if you want to play around.